% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{huge_testing_data}
\alias{huge_testing_data}
\title{Huge testing dataset}
\format{
A list with a data frame and a numeric vector. The data frame has 1000 rows
and 51 variables. The vector has length 3.
}
\usage{
data(huge_testing_data)
}
\description{
List of a dataset and the predictor variables used to generate the
response variable in the dataset. The dataset was made for testing \code{select}
on a dataset with a large number of variables. The response variable is in
the 51st column and is generated using \code{X15}, \code{X26}, and \code{X45}.
}
\keyword{datasets}
