% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{large_testing_data}
\alias{large_testing_data}
\title{Large testing dataset}
\format{
A data frame with 1000 rows and 14 variables.
}
\usage{
data(large_testing_data)
}
\description{
Dataset made for testing \code{select} on a dataset with
a limited number of variables. The response variable is in
the 14th column and is generated using \code{x5}, \code{x8}, and \code{x13}.
}
\keyword{datasets}
