% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select.R
\name{select}
\alias{select}
\title{Select Method of Genetic Algorithm}
\usage{
select(
  x,
  reg,
  reg_fun,
  family = "gaussian",
  fun = AIC,
  ngen = 40,
  sel = gen_selection,
  cross = gen_crossover,
  mut = gen_mutate
)
}
\arguments{
\item{x}{a data frame}

\item{reg}{regression type}

\item{reg_fun}{regression formula}

\item{fun}{fitness function. the default is AIC.}

\item{ngen}{max number of generations. the default is 40.}

\item{sel}{generation selection function. a default function is provided, but the user may choose to use a different method.}

\item{cross}{crossover function. a default function is provided, but the user may choose to use a different method.}

\item{mut}{mutate function. a default function is provided, but the user may choose to use a different method.}
}
\value{
Principal components and restricted data set.
}
\description{
Variable selection using genetic algorithm.
}
\examples{

x_small <- testdata(nvar = 6, rper = 0.5, npoints = 100, noisevar = 1, linkfunc = identity, dist = 'Gaussian', -10, 10)
x_large <- testdata(nvar = 20, rper = 0.5, npoints = 100, noisevar = 1, linkfunc = identity, dist = 'Gaussian', -10, 10)

reg <- glm

reg_fun_small <- Y ~ X1 + X2 + X3 + X4 + X5 + X6
reg_fun_large <- Y ~ X1 + ... + X20

family <- "gaussian"
fun <- AIC

select(x, reg, reg_fun, family, fun)
}
